make the random color generation be of type float32

+++ train_nerf_face_simplified.py
@@ -250,7 +250,7 @@ def main():
         if hasattr(cfg.dataset, "mask_face") and cfg.dataset.mask_face:
             out = face_seg_net.infer(img_target).astype(np.float32)
             # Mask out image and make background random
-            img_target = img_target*out[:,:,None]+((1-out[:,:, None])*np.random.uniform(0,1,(1,1,3)))
+            img_target = img_target*out[:,:,None]+((1-out[:,:, None])*np.random.uniform(0,1,(1,1,3))).astype(np.float32)
         img_target = img_target.to(device)
